requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
beartype==0.13.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fast-bitrix24==1.5.16
frozenlist==1.3.3
html5lib==1.1
icontract==2.6.2
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.5.0
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pypdf==3.7.1
pypng==0.20220715.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.9
yarl==1.8.2
python-telegram-bot==13.14
